loader.efi: main comparison between signed and unsigned integer expressions
../main.c: In function 'main':
../main.c:674:16: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  for (i = 0; i < sz / sizeof(boot_order[0]); i++)
                ^
cc1: all warnings being treated as errors

+++ usr/src/boot/sys/boot/efi/loader/main.c
@@ -693,7 +693,7 @@ main(int argc, CHAR16 *argv[])
 	sz = sizeof (boot_order);
 	efi_global_getenv("BootOrder", &boot_order, &sz);
 	printf("   BootOrder:");
-	for (i = 0; i < sz / sizeof (boot_order[0]); i++)
+	for (i = 0; i < (int)(sz / sizeof (boot_order[0])); i++)
 		printf(" %04x%s", boot_order[i],
 		    boot_order[i] == boot_current ? "[*]" : "");
 	printf("\n");
